+++ capture/engine_scripts/playwright/onBefore.js
@@ -1,0 +1,3 @@
+module.exports = async (page, scenario, viewport, isReference, browserContext) => {
+  await require('./loadCookies')(browserContext, scenario);
+};