+++ core/util/extendConfig.js
@@ -23,6 +23,7 @@ function extendConfig (config, userConfig) {
   config.asyncCompareLimit = userConfig.asyncCompareLimit;
   config.backstopVersion = version;
   config.dockerCommandTemplate = userConfig.dockerCommandTemplate;
+  config.scenarioLogsInReports = userConfig.scenarioLogsInReports;
   config.customReports = userConfig.customReports;
   return config;
 }